requirements.txt
absl-py==1.4.0
accelerate==0.18.0
adversarial-robustness-toolbox==1.14.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
blosc2==2.0.0
boto3==1.26.94
botocore==1.29.94
braceexpand==0.1.7
brambox==4.3.0
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.26.0
comm==0.1.2
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
cox==0.1.post3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.35
datasets==2.12.0
debugpy==1.6.6
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.14.0
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fairscale==0.4.4
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
geotorch==0.3.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.51.3
huggingface-hub==0.14.1
pip-install==1.3.5
identify==2.5.24
idna==2.10
imageio==2.28.1
importlib-metadata==6.1.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.13
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
lit==15.0.7
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.6
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.1
plotly==5.14.1
portalocker==2.7.0
pre-commit==3.3.1
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytorch-pretrained-gans @ git+https://github.com/lukemelas/pytorch-pretrained-gans@2982fdab4e683165e45bc2f4a64c2942a7a3a1b7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
regex==2023.3.23
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
robustness==1.2.1.post2
rsa==4.9
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.19.1
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
streamlit==1.22.0
sympy==1.11.1
tables==3.8.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.4.0
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
tzlocal==5.0.1
ultralytics==8.0.101
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
virtualenv==20.23.0
wandb==0.14.2
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
